move export to trait
